fix(vault): tighten BTC tx resolution and add mempool link to collateral
Address PR review:
- resolveDisplayTx now rejects empty / "0x" peginTxHash explicitly so
  Deposit rows surface a real indexer issue instead of silently labelling
  themselves ETH.
- CollateralVaultItem links the peg-in hash to mempool, matching the
  Activity tab; both call sites share a new utils/explorer helper.
